refactor: move intervals to contract_interface crate
